requirements.txt
# Django
Django==3.2

# Django Rest Framework
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
django-filter==2.4.0

# Database
psycopg2-binary==2.8.6
dj-database-url==0.5.0

# Extensions
django_extensions==3.1.2
ipython==7.22.0

# HTTP
requests==2.25.1

# Testing
coverage==5.5
pytest==6.2.3
pytest-django==4.2.0
pytest-cov==2.11.1

# Production server
uvicorn==0.13.4

# Sentry
sentry-sdk==1.0.0
